requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

numpy<2
torch==2.*            # matches your CUDA 12.1 module
sentence-transformers
faiss-gpu-cu12==1.8.0 # or faiss-cpu==1.8.0 if you don't need GPU FAISS

arxiv
feedparser
requests
tqdm
pyyaml
orjson
ujson

